backend/pyproject.toml
[tool.poetry]
name = "infinipaper-backend"
version = "0.1.0"
description = "InfiniPaper backend (FastAPI + SQLModel + Postgres + pgvector)"
authors = ["You <you@example.com>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.114.0"
uvicorn = {extras = ["standard"], version = "^0.30.6"}
sqlmodel = "^0.0.21"
sqlalchemy = "^2.0.34"
psycopg = {extras = ["binary"], version = "^3.2.1"}
pydantic-settings = "^2.4.0"
python-dotenv = "^1.0.1"
alembic = "^1.13.2"
celery = "^5.4.0"
redis = "^5.0.8"
httpx = "^0.27.2"
loguru = "^0.7.2"
pgvector = "^0.2.5"
python-multipart = "^0.0.9"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
ruff = "^0.5.7"
black = "^24.8.0"

[tool.ruff]
line-length = 100

[tool.black]
line-length = 100

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"